+++ .tape.js
@@ -4,7 +4,7 @@ module.exports = {
 			message: 'supports basic usage'
 		},
 		'basic:base64': {
-			message: 'supports { utf8: false } usage',
+			message: 'supports { "utf8": false } usage',
 			options: {
 				utf8: false
 			}